fix(bootstrap): same-account modern bootstrapping still requires policy ARNs (#9867)
The current bootstrapping experience has sharp edges. It requires you
to pass `--cloudformation-execution-policies` in all cases, even if
you just want a "simple", same-account bootstrap stack. If you bootstrap
your own accounts, you probably don't have a centralized SecOps
team that is limiting you, and you're probably intending to deploy
everything using the CDK, which means you're looking for AdministratorAccess.

In effect, you are forced to constantly Google and copy/paste the ARN
for `AdministratorAccess` (because who can remember that)
and it's a bad experience.

In the case of bootstrapping an account just for use "by itself",
however (which is the 90% use case for bootstrapping),
the trust boundary is very clear and there's no risk of privilege
escalation between accounts.

Add an optimization where in the case of a "simple", non-cross account
bootstrap, we'll default to the AdministratorAccess ARN for you.

Once you establish `--trust` with another account, we'll still
force you to spell out the execution policy you'll want to use though.

Fixes #8571.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*