chore(synthetics): make runtime a required property (#11163)
The canary runtime property determines which [synthetics runtime version](https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_Library.html) will be included in the lambda function created by the Synthetics service. The difference between runtime versions could be as dramatic as a change to the supported nodejs version.  Given that the "sane default" value changes as new runtime versions are released, it is better to leave it to the user to make an informed decision which version to use.

Merged the previous two integration tests so it will be easier to test. 

BREAKING CHANGE: `runtime` is now a required property.  
 


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*