feat(efs): import access point - `fromAccessPointAttributes()` (#10712)
Cannot use `efs.AccessPoint.fromAccessPointId()` with `lambda.FileSystem.fromEfsAccessPoint()`. the former returns an `IAccessPoint` when the later expect an `AccessPoint`. I think following the CDK guidelines, `lambda.FileSystem.fromEfsAccessPoint()` should expect an `IAccessPoint`, not an `AccessPoint`.

Argument of type `IAccessPoint` is not assignable to parameter of type `AccessPoint`.

### Solution
----

Add a new import method to the `AccessPoint` class called `fromAccessPointAttributes()` allowing to pass a fileSystem as an attribute.

Closes #10711.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*